requirements.txt
Django==1.7
argparse==1.2.1
bottle==0.12.7
bottle-session==0.4
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.1.1
httplib2==0.9
psycopg2==2.5.4
python-instagram==1.1.3
redis==2.10.3
simplejson==3.6.4
sqlparse==0.1.13
static3==0.5.1
wsgiref==0.1.2
